Pilko KoealustaMappingService kapeampiin vastuisiin
KoealustaMappingService oli 543 rivinen, koska se hoiti Moodle-vastauksen
jäsentämisen, validoinnin, ONR-virheiden tulkinnan sekä virheiden
muunnoksen rinnakkain. Erotetaan nämä eri tiedostoihin:

- KoealustaResponseParser: Moodle-JSONin parsinta ja MoodleException
- KoealustaSuoritusValidator: toOppija, completionToEntity ja validate-
  apurit
- KoealustaMappingError: sealed-hierarkia siirretty top-leveliksi (vanha
  nimi Error varjosti kotlin.Erroria)
- oppijanumero/OppijanumerorekisteriDebugInfo: ONR-vastauksen tyypilliset
  virheet ja HTTP-statusten suomennokset
- oppijanumero/OppijanumeroTroubleshooting: troubleshootOppijanumero
  laajennusfunktio OppijanumeroTroubleshootingServicelle

KoealustaMappingService jää orkesteroimaan parserin, validaattorin ja
ONR-palvelut convertToEntityyn sekä muuntamaan virheet
KielitestiSuoritusErroriksi.

Lisäksi siivottu käyttämätön validateNonEmpty-funktio ja korvattu
unicode-escape-merkit kirjaimilla validointistringeissä.
